Refactor: Move and rename vazkii.botania.common.lib.LibItemAges to vazkii.botania.common.core.helper.ItemAgeHelper
