+++ src/main/java/vazkii/botania/common/block/subtile/functional/SubTilePollidisiac.java
@@ -22,7 +22,7 @@ import vazkii.botania.api.subtile.RadiusDescriptor;
 import vazkii.botania.api.subtile.SubTileFunctional;
 import vazkii.botania.common.core.handler.MethodHandles;
 import vazkii.botania.common.lexicon.LexiconData;
-import vazkii.botania.common.lib.LibItemAges;
+import vazkii.botania.common.core.helper.ItemAgeHelper;
 import vazkii.botania.common.lib.LibObfuscation;
 
 public class SubTilePollidisiac extends SubTileFunctional {
@@ -53,7 +53,7 @@ public class SubTilePollidisiac extends SubTileFunctional {
 							continue;
 						}
 
-						if(!LibItemAges.isLegalForFunctional(age, slowdown) || item.isDead)
+						if(!ItemAgeHelper.isLegalForFunctional(age, slowdown) || item.isDead)
 							continue;
 
 						ItemStack stack = item.getEntityItem();
