+++ src/main/java/vazkii/botania/common/block/subtile/functional/SubTileRannuncarpus.java
@@ -46,7 +46,7 @@ import vazkii.botania.common.core.handler.ConfigHandler;
 import vazkii.botania.common.core.handler.MethodHandles;
 import vazkii.botania.common.item.block.ItemBlockSpecialFlower;
 import vazkii.botania.common.lexicon.LexiconData;
-import vazkii.botania.common.lib.LibItemAges;
+import vazkii.botania.common.core.helper.ItemAgeHelper;
 import vazkii.botania.common.lib.LibObfuscation;
 
 public class SubTileRannuncarpus extends SubTileFunctional {
@@ -89,7 +89,7 @@ public class SubTileRannuncarpus extends SubTileFunctional {
 					continue;
 				}
 
-				if(!LibItemAges.isLegalForFunctional(age, slowdown) || item.isDead || item.getEntityItem().isEmpty())
+				if(!ItemAgeHelper.isLegalForFunctional(age, slowdown) || item.isDead || item.getEntityItem().isEmpty())
 					continue;
 
 				ItemStack stack = item.getEntityItem();
