+++ src/main/java/vazkii/botania/common/block/tile/TileOpenCrate.java
@@ -23,7 +23,7 @@ import net.minecraft.world.World;
 import vazkii.botania.api.state.BotaniaStateProps;
 import vazkii.botania.common.block.ModBlocks;
 import vazkii.botania.common.core.handler.MethodHandles;
-import vazkii.botania.common.lib.LibItemAges;
+import vazkii.botania.common.core.helper.ItemAgeHelper;
 
 public class TileOpenCrate extends TileSimpleInventory {
 
@@ -76,7 +76,7 @@ public class TileOpenCrate extends TileSimpleInventory {
 
 		if(redstone) {
 			try {
-				MethodHandles.itemAge_setter.invokeExact(item, LibItemAges.OPEN_CRATE_REDSTONE_AGE);
+				MethodHandles.itemAge_setter.invokeExact(item, ItemAgeHelper.OPEN_CRATE_REDSTONE_AGE);
 			} catch (Throwable ignored) {}
 		}
